finalisation for the alpha version. please playtest!
remove leaderboard and add a nice landing page.
add admin contact to config so people can reach the sysadmin.
add the ability to change your display name ingame through
the abuse of the contract system. usernames will not be saved
across server restarts, but can be changed anytime.
reword contract scheduling for reasons
due to unforseen concequences, we now too parse the protobuf
on boot from the base64 representation. the tool to make it
human readable stays. but dont use its output files, they
are broken for parsing. (as the protojson page warned)